+++ Vault/src/utopia/vault/nosql/targeting/many/AccessMany.scala
@@ -6,7 +6,10 @@ import utopia.flow.util.Mutate
 import utopia.vault.model.enumeration.SelectTarget
 import utopia.vault.model.immutable.Table
 import utopia.vault.model.mutable.ResultStream
+import utopia.vault.model.template.HasTable
 import utopia.vault.nosql.factory.FromResultFactory
+import utopia.vault.nosql.read.DbReader
+import utopia.vault.nosql.read.parse.ParseResultStream
 import utopia.vault.nosql.template.Deprecatable
 import utopia.vault.sql.{Condition, OrderBy, SqlSegment, SqlTarget}
 
@@ -14,33 +17,34 @@ object AccessMany
 {
 	// OTHER    ------------------------------
 	
-	def apply[A](factory: FromResultFactory[A]): AccessMany[A] = apply(factory, useDefaultOrdering = false)
+	def apply[A](factory: DbReader[Seq[A]] with HasTable): AccessMany[A] =
+		_AccessMany[A](factory.target, factory.table, factory.selectTarget, factory)
 	def apply[A](factory: FromResultFactory[A], useDefaultOrdering: Boolean): AccessMany[A] =
-		_AccessMany[A](factory.target, factory.table, factory.selectTarget, s => factory(s.buffer),
+		_AccessMany[A](factory.target, factory.table, factory.selectTarget, factory,
 			ordering = if (useDefaultOrdering) factory.defaultOrdering else None)
 	
 	def apply[A](target: SqlTarget, table: Table, selectTarget: SelectTarget, condition: Option[Condition] = None,
 	             ordering: Option[OrderBy] = None, prepare: Mutate[SqlSegment] = Identity)
-	            (f: ResultStream => Seq[A]): AccessMany[A] =
-		_AccessMany[A](target, table, selectTarget, f, condition, ordering, prepare)
+	            (parse: ParseResultStream[Seq[A]]): AccessMany[A] =
+		_AccessMany[A](target, table, selectTarget, parse, condition, ordering, prepare)
 	
 	def table[A](table: Table, condition: Option[Condition] = None, ordering: Option[OrderBy] = None,
 	             prepare: Mutate[SqlSegment] = Identity)
-	            (f: ResultStream => Seq[A]) =
-		apply[A](table, table, SelectTarget.table(table), condition, ordering, prepare)(f)
+	            (parse: ParseResultStream[Seq[A]]) =
+		apply[A](table, table, SelectTarget.table(table), condition, ordering, prepare)(parse)
 	
-	def tables[A](first: Table, second: Table, more: Table*)(f: ResultStream => Seq[A]): AccessMany[A] =
+	def tables[A](first: Table, second: Table, more: Table*)(parse: ParseResultStream[Seq[A]]): AccessMany[A] =
 		apply[A](more.foldLeft(first join second) { _ join _ }, first,
-			SelectTarget.tables(Pair(first, second) ++ more))(f)
+			SelectTarget.tables(Pair(first, second) ++ more))(parse)
 	
-	def active[A](factory: FromResultFactory[A] with Deprecatable): AccessMany[A] =
+	def active[A](factory: DbReader[Seq[A]] with HasTable with Deprecatable): AccessMany[A] =
 		apply(factory).filter(factory.nonDeprecatedCondition)
 	
 	
 	// NESTED   ------------------------------
 	
 	private case class _AccessMany[+A](target: SqlTarget, table: Table, selectTarget: SelectTarget,
-	                                   f: ResultStream => Seq[A],
+	                                   f: ParseResultStream[Seq[A]],
 	                                   accessCondition: Option[Condition] = None, ordering: Option[OrderBy] = None,
 	                                   prepare: Mutate[SqlSegment] = Identity)
 		extends ConcreteAccessManyLike[A, AccessMany[A]] with AccessMany[A]